Automatic updates
Implemented automatic updates
The app check for updates when it starts with a configurable delay (now
5m). After that, the app check for updates every 2 hours (this is
configurable too).
If the app detects an update, it is downloaded and the app is silently
rebooted when there aren't active transfers.
All the information about the update is hashed and signed using a
private key. If the update information or any of the updated files can't
be verified, the update is aborted. The app makes a rollback if the
update can't be applied. To allow testing, an example public and private
keys have been placed in the code.
The update system check every updated file to know if it's already
installed (the file hasn't changed in the update) of it's already
downloaded (in a previous failed update).
To support the silent reboot, the management of the thread that receives
the messages from the shell extension has been enhanced to allow a clean
exit and prevent hangs when the app is closed.
The code of the main class (MegaApplication) has been commented and some
useless functions have been removed.