Check for dupplicate sync names
Dialog box to ask users for a new sync identifier if the local folder
name is used by other synced folder.
Fix checking for path states
Used the sync name to identify syncs in the GUI and to create shortcuts
in the "Favourites" folder, instead of using the local folder name.