Several changes:
Changed the initial setup wizard to allow the selection of the two new
usage modes: "Full account sync" and "Selective sync".
Show a warning if the size of a new synced folder (in MEGA) is more than
2 GB.
Changed the behavior of the "Syncs" button if the root folder is synced.
Faster conversion from QString to wchar_t *.
Bug fixes.