Get public links from the shell extension
The new menu item appears when there are selected files already synced.
If there are synced and unsynced files selected, both menu items are
shown (upload and get link).
The menuitem text now contains the number of files and folders that are
affected by the operation.
The new feature is fully implemented, the app receives the new messages
from the DLL and it completes the requests.