New way to access nodes
Changed the way in which nodes are used in the upper layer for the SDK,
and in the app. Now, the SDK thread never waits for the GUI thread and
the SDK can delete nodes at any moment (for example, external deletions
from other devices) without crashing the app.
Fixed memory leaks using getNodePath.
Other bug fixes.