Minor changes in the setup wizard
Clear user information from the "Create accout" page when the account is
created of the user clicks the back button
Propagate the email from the "Create account" to the "Login" page after
a successful account creation.
Clear the password field when users go back to the "Login" page or click
the "back" button in that page.