Fixed a bug changing the local folder associated with a MEGA folder
To change the local folder associated with a MEGA folder, the user has
to delete the sync and to add it again. If the user didn't apply the
changes between these operations, the local folder wasn't correctly
changed and was restored in the next restart. This patch fixes the bug.