New behavior and design for the systray menu
It's shown when the systray icon is clicked with the middle mouse button
or when the settings button is clicked,
Removed the "Pause" action
New menu item that shows the version of MEGAsync when it's updated and
"Install update" when there is an available update.
Show the information popup in the corner next to the cursor (instead of
always in the right-bottom corner).