For now, disable only recreating InfoDialog on screen config difference
The scenario where optimising the InfoDialog recreation that doesn't work for me occurs in normal usage - I lock the laptop, go to lunch, come back and unlock the laptop, and now clicking the taskbar icon shows a wrongly-sized InfoDialog.    My setup has 3 monitors:  1. laptop monitor, 200%.  2. external 4k monitior, 100% (set to main display). 3. external 4k monitor, 300% (for testing).    To simulate the lunch break, I lock the laptop, unplug the screens, wait 30 seconds, plug them back in, and unlock the laptop.    So at unlocking time, in fact the screens are back in the same configuration they were the last time we checked so we don't detect a difference.  However QT has somehow internally done something to the dialog and/or widget scaling factors, which seems extremely difficult to check or influence.  The result is the dialog contents at 200%, but confined in a 100% size window so we can only see 1/4 of it, and can't access the cog menu etc.