Only lock the sdkMutex once during transfer view painting
We used to lock it many times, and get blocked a lot by the SDK.  Just locking it once (only if needed) and unlocking it at the end of the painting operation makes the GUI much more responsive during busy upload activity.
Currently needs the lockMutex() and unlockMutex() which are being added to the SDK, a temporary macro indicates if those are available.