Modifed the megasync.show logic so that we don't end up starting megasync if the original instance showed the dialog.
The file has a few bytes added, to request the original instance to show the dialog.   If the new instance sees the file size go back to 0, that means the original instance has done so.
In that case, the new instance exits immediately.
In the meantime, the new instance checks periodically if it can get the unique lock, and if it can then it does start up, and erases the content of that file for a fresh start.