Change when the local SSL Certs are requested
The login request aborts all other requests in the queue, so the first
local SSL Certs request is made after the login request finishes,
instead of being queued together.
If the login request silently aborts the SSL Cert request, the local
https server does not start as the request seems to be pending.