Log performance and related improvements
Tested unicode paths by settings "logs" folder to be a username from a user bug report from a while back, where just using wide characters was not enough.
5 clicks on user avatar in settings to log to desktop is working
Bug report upload is working
Performance is much better, the ultra simple solution was showing in 15% of profiler samples when busy downloading, now down to 0.75%
In my testing, there were a lot of duplicate log lines from cURL, sometimes repeated hundreds of times.  Those are now detected and replaced with eg.
	01/26-05:30:02.915 29556 DTL  cURL DEBUG: schannel: failed to decrypt data, need more data [net.cpp:1705]
	[repeated x142]
Columns are adjusted to line up nicely, with fixed size for ERR, WARN etc and the file and line at the end, so the main text is easy to read.
The source filename reported was coming out as its full path, and its text was about 50% of the log content so I've adjusted that to just be the leaf name, and pre-calculated at compile time with `constexpr`.
With those changes the 10MB main log compresses to 850KB instead of 450KB, so I've adjusted the number of files to 50 (was 100) to use the same storage.