Fix for more timers being created frequently when only uploading or only downloading.
InfoDialog::updateTransfersCount() is usually called when any notification for transfers occurs.  It would create a new 5 second timer if there are no uploads, or if there are no downloads
This fix just prevents another timer if there is already one running for that case.
Also a small adjustment for the CMakeLists, MEGAsync uses the performance logging, so turn that on by default.