Fixes for windows build scripts after trying the recent changes on the real build machine
Don't erase existing folders.  Eg if x64 built but x86 had a problem, we don't want to go back to square one and rebuild x64.   The full_build script can do that as it authors desire that, the separate step scripts should not.  It's very easy to manually remove those if needed, and confirm by eyeball.

Don't try to set CORES.  This was causing a syntax error when not set.  And, the build machine is already using 4 cores at high CPU anyway.