Move step panel string modification from MainFlow to ResumePageForm
This change is made to allow the strings in steps 3 and 4 to always be translatable.
The first time they were translatable, but after they were assigned in MainFlow for the ResumePage, their connections to translate were broken.
Also, this change aligns the use of the stepTexts as in syncs and backups.